refactor(core): extract context-budget from memory.ts
Move calculateContextBudget() and buildContextWithMemory() from
memory.ts into a new context-budget.ts module. This fixes a conceptual
cycle where the memory data layer was importing from the agent
orchestration layer. The new module uses type-only imports for
ContextStats, Memory, and MemoryStore to avoid runtime circular
dependencies.

Test files updated to import from the new module path.
